Block reboots unless new version is permitted (#1)
* Block reboots unless new version is permitted

If `NEW_VERSION` is less than $DESIRED_VERSION then reset the
update-engine status and check for a newer version.

Allow reboot if `NEW_VERSION == $DESIRED_VERSION`.

Always allow a reboot if `$DESIRED_VERSION == "current"`. This provides
the possibility of running the DS across all envs for the purposes of
parity while still unpinning some envs. It also gives us a quick switch
to block reboots if we find an issue.

* Address comments

- Fix command splitting in Dockerfile so that it adheres to shell
  guidelines
- Rename `okay` in annotation name to `ok`